requirements.txt
pandas>=2.0.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
python-dotenv>=1.0.0

# Required for huggingface endpoint support
langchain-community>=0.1.0
huggingface-hub>=0.13.0
transformers>=4.30.0

